Project 0x11 C - printf

+++ holberton.h
@@ -6,7 +6,7 @@ int _putchar(char c);
 int print_str(char *c);
 int print_int(long int num);
 int print_Str(char *c);
-int print_bin(unsigned long int num);
+int print_bin(unsigned int num);
 int print_hex(unsigned long int num);
 int print_hexlow(unsigned long int num);
 int print_oct(unsigned long int num);
